requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
plotly
python-dotenv
